graph/requirements.txt
pandas
influxdb-client
matplotlib
seaborn
pillow